Rename constants to types
