+++ app/actions/topics.js
@@ -2,7 +2,7 @@
 import { polyfill } from 'es6-promise';
 import request from 'axios';
 import md5 from 'spark-md5';
-import * as types from 'constants/index';
+import * as types from 'types';
 
 polyfill();
