+++ app/actions/users.js
@@ -2,7 +2,7 @@ import { polyfill } from 'es6-promise';
 import request from 'axios';
 import { push } from 'react-router-redux';
 
-import * as types from 'constants';
+import * as types from 'types';
 
 polyfill();
