+++ app/tests/actions/topics-test.js
@@ -6,7 +6,7 @@ import { polyfill } from 'es6-promise';
 import axios from 'axios';
 import expect from 'expect';
 import * as actions from 'actions/topics';
-import * as types from 'constants';
+import * as types from 'types';
 
 polyfill();
