+++ app/tests/reducers/topic-test.js
@@ -1,7 +1,7 @@
 import expect from 'expect';
 import md5 from 'spark-md5';
 import reducer from 'reducers/topic';
-import * as types from 'constants';
+import * as types from 'types';
 
 describe('Topics reducer', () => {
   it('Should return the initial state', () => {
